requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
asyncer==0.0.1
asynctest==0.13.0
attrs==22.1.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
fastapi==0.80.0
filelock==3.8.0
filetype==1.1.0
firebase-admin==5.3.0
Flask==2.2.2
flatbuffers==2.0.7
frozenlist==1.3.1
gdown==4.5.1
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
h11==0.13.0
httplib2==0.20.4
humanfriendly==10.0
idna==3.3
ImageHash==4.2.1
imageio==2.21.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.39.0
MarkupSafe==2.1.1
mpmath==1.2.1
msgpack==1.0.4
multidict==6.0.2
networkx==2.6.3
numba==0.56.0
numpy==1.21.6
onnxruntime==1.12.1
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.1
PyMatting==1.1.8
pyparsing==3.0.9
PySocks==1.7.1
python-multipart==0.0.5
PyWavelets==1.3.0
rembg==2.0.24
requests==2.28.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
sympy==1.10.1
tifffile==2021.11.2
tqdm==4.64.0
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
watchdog==2.1.9
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
